+++ with-tailwindcss-app/package.json
@@ -15,6 +15,7 @@
     "next": "latest",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-hook-form": "^7.41.0",
     "swr": "^2.0.0"
   },
   "devDependencies": {
