+++ with-tailwindcss-app/pages/components/SuppliersPage/SupplierCard.tsx
