+++ with-tailwindcss-app/pages/components/common/StyledModal.tsx
